Prototype of frame "glass" approach
A frame can temporarily expand to allow "popup" content from within it to
expand outside of its visual bounds.

A "cover" system is used in the host so that non-popup areas can receive
clicks that will be handled by the host.

To lessen the risk of clickjacking, popup content could have mandatory "chrome"
that makes it look like a popup element anchored to the frame, so that it does't
impersonate arbitrary UI.